Ads and blank message files, plus housekeeping.
We don't need a separate license file for this as it is part of the
web site, which displays the license in the footer.

Made the README nicer.

.gitignored stuff (copied over from gcd-django/.gitignore, minus
the python-related parts).

The HTML files in this commit are unchanged from their form in gcd-django.
They will generally be edited by tech admins and don't need instructions.